Fix pagination URL encoding.
APIs supporting pagination use a square bracket parameter to index the page and
the offset. The format is not correctly encoded by the urlencode method from
urllib when building the URL parameter list, since multidimensional dictionaries
are not supported.

For following example:

{'longitude': 0.1278, 'latitude': 51.5074, 'page': {'offset': 10}}

is encoded as:

longitude=0.1278&latitude=51.5074&page={offset : 10}

instead of:

longitude=0.1278&latitude=51.5074&page[offset]=10

The new private method _urlencode parses and prepares the dictionary before the
actual call to urlencode is done.

Resolves: #28