removed duplicate readme entries, tests, rewrote test case
this removes a duplicate readme entry. This also renames
the readme entry to be flight seatmaps instead of just
seatmaps to fit more with the style of the readme

the testcase for seatmaps.get was changed to pass in the
args in the style that is also used in the readme. the
reason the args are passed in like this is because there
is a hyphen present in the arg name.

this also removes a duplicate post test.