Sprutner/add seatmap explorer (#60)
* Adds seatmap explorer to the python-sdk

This adds functionality to the amadeus python SDK to allow
use of the "SeatMap Display" API endpoint within the
python SDK.

* Adds seatmap explorer to the python-sdk

This adds functionality to the amadeus python SDK to allow
use of the "SeatMap Display" API endpoint within the
python SDK.

Adds documentation and fixed URL

* added tests for seatmap

adds basic tests for both the get and post functions
for seatmap explorer

* updated docs on seatmap explorer

made the documentation more specific, and removed the invalid
offerid, as the get method only supports order id from flights
that have already been ordered.

* added more tests

added required tests on paths and .get methods

* fixed docstring typo

* Adds seatmap explorer to the python-sdk

This adds functionality to the amadeus python SDK to allow
use of the "SeatMap Display" API endpoint within the
python SDK.

Adds documentation and fixed URL

* added tests for seatmap

adds basic tests for both the get and post functions
for seatmap explorer

* fix issues from github merge conflict tool

* removed duplicate readme entries, tests, rewrote test case

this removes a duplicate readme entry. This also renames
the readme entry to be flight seatmaps instead of just
seatmaps to fit more with the style of the readme

the testcase for seatmaps.get was changed to pass in the
args in the style that is also used in the readme. the
reason the args are passed in like this is because there
is a hyphen present in the arg name.

this also removes a duplicate post test.

Co-authored-by: Anna Tsolakou <tsolakou.a@gmail.com>